requirements.txt
android-backup==0.2.0
appdirs==1.4.4
attrs==20.3.0
cffi==1.14.4
click==7.1.2
construct==2.10.58
croniter==0.3.37
cryptography==3.3.1
ifaddr==0.1.7
natsort==7.1.1
netifaces==0.10.9
paho-mqtt==1.5.1
pycparser==2.20
python-dateutil==2.8.1
python-miio==0.5.4
pytz==2021.1
schedule==1.0.0
six==1.15.0
tqdm==4.56.0
zeroconf==0.28.8
